requirements.txt
Flask-Cors>=3.0.10
Flask-Script>=2.0.6
Flask>=2.0.2
PyJWT>=2.3.0
marshmallow>=3.14.0
openpyxl>=3.0.9
attrs>=21.4.0
gunicorn>=20.1.0
sendgrid>=6.9.6
